alter the authentication to be more "reactive"
Now a modal dialog is put up that asks the user whether they want to continue, when the token is about to expire. And if they do not answer at that point they will be automatically logged out.

In addition, an activity timer is used to avoid the prompt.

This fixes #24